% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_preprocessing.R
\name{detect_languages}
\alias{detect_languages}
\title{Detect Languages with fastText}
\usage{
detect_languages(
  x,
  text_col = "text",
  id_col = NULL,
  lang_guess_col = NULL,
  model_path = NULL,
  prob_threshold = 0.25,
  und_label = "und",
  max_char = 5000,
  tokenize_sentences = FALSE,
  threads = parallel::detectCores(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Character vector or data.frame containing texts to clean.}

\item{text_col}{Character, name of the text column if \code{x} is a data.frame.
Ignored if \code{x} is a character vector. Default = "text".}

\item{id_col}{Character, optional column name in the input data.frame to
preserve as \code{id}. Default = NULL.}

\item{lang_guess_col}{Character, optional column name in the input data.frame
to preserve as \code{lang_guess}. Default = NULL.}

\item{model_path}{Character, path to the fastText \code{lid.176.bin} model.
Defaults to \verb{~/.cache/easieRnmt/lid.176.bin}.}

\item{prob_threshold}{Numeric, threshold below which the detected language is replaced
by \code{und_label} (or by the user-provided \code{lang_guess}). Default = 0.25.}

\item{und_label}{Character, label for undefined language if confidence is below
\code{prob_threshold} and no \code{lang_guess} column is provided. Default = "und".}

\item{max_char}{Integer, maximum number of characters per text. Texts longer
than this are truncated. Default = 5000.}

\item{tokenize_sentences}{Logical, whether to split texts into sentences and
chunks. Default = TRUE.}

\item{threads}{Integer, number of threads for fastText. Default = \code{parallel::detectCores()}.}

\item{verbose}{Logical, whether to print progress messages. Default = TRUE.}

\item{...}{Additional parameters passed on to \code{clean_text()}.}
}
\value{
A data.table with standardized columns:
\code{sen_id}, \code{doc_idx}, \code{sen_idx}, optional \code{id}, \code{text_orig},
\code{text_clean}, optional \code{lang_guess}, \code{lang}, and \code{lang_prob}.
}
\description{
Identifies the language of each text in a character vector or data.frame
using fastText's pretrained language identification model. Calls \code{clean_text()}
internally to normalize and optionally tokenize the text before prediction.
}
