% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_installs.R
\name{install_torch}
\alias{install_torch}
\title{Install PyTorch in a Conda Environment}
\usage{
install_torch(conda_env_name = "r-easynmt", conda_path = NULL, verbose = TRUE)
}
\arguments{
\item{conda_env_name}{Character. Name of the conda environment. Default "r-easynmt".}

\item{conda_path}{Character. Path to Conda installation. If NULL, autodetect.}

\item{verbose}{Logical. If TRUE, prints progress messages. Default TRUE.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Installs PyTorch and related libraries (\code{torch}, \code{torchvision}, \code{torchaudio})
in a Conda environment created with \code{\link[=install_conda_env]{install_conda_env()}}. The function detects CUDA/GPU
availability and installs the appropriate build. If the GPU build URL is invalid, it falls back to CPU.
}
